// src/types/User.ts
export type UserProps = {
  login: string;
  name: string;
  avatar_url: string;
  bio: string;
  company: string;
  email: string;
  location: string;
  html_url: string;
  followers: number;
  followers_url: string;
  following: number;
  following_url: string;
  repos_url: string;
};